+++ src/exactly_lib/help/entities/syntax_elements/objects/type_list.py
@@ -6,6 +6,7 @@ from exactly_lib.definitions.cross_ref.app_cross_ref import SeeAlsoTarget
 from exactly_lib.definitions.cross_ref.name_and_cross_ref import cross_reference_id_list
 from exactly_lib.definitions.entity import syntax_elements, types, concepts
 from exactly_lib.definitions.test_case.instructions import define_symbol
+from exactly_lib.help.entities.syntax_elements.common.list_syntax import list_syntax_description
 from exactly_lib.help.entities.syntax_elements.contents_structure import SyntaxElementDocumentation
 from exactly_lib.util.cli_syntax.elements import argument as a
 from exactly_lib.util.textformat.structure.core import ParagraphItem
@@ -32,7 +33,7 @@ class _Documentation(SyntaxElementDocumentation):
         ]
 
     def main_description_rest_paragraphs(self) -> List[ParagraphItem]:
-        return []
+        return list_syntax_description(types.LIST_TYPE_INFO.name)
 
     def see_also_targets(self) -> List[SeeAlsoTarget]:
         info_refs = cross_reference_id_list([