+++ src/exactly_lib/impls/types/program/parse/parse_arguments.py
@@ -1,5 +1,7 @@
+import functools
+
 from exactly_lib.impls.file_properties import FileType
-from exactly_lib.impls.types.parse import parse_list
+from exactly_lib.impls.types.list_ import parse_list, generic_parser
 from exactly_lib.impls.types.path import parse_path
 from exactly_lib.impls.types.program import syntax_elements
 from exactly_lib.impls.types.program.command import arguments_sdvs
@@ -8,7 +10,7 @@ from exactly_lib.section_document.element_parsers import token_stream_parsing as
 from exactly_lib.section_document.element_parsers.instruction_parser_exceptions import \
     SingleInstructionInvalidArgumentException
 from exactly_lib.section_document.element_parsers.ps_or_tp.parsers import Parser, ParserFromTokenParserBase
-from exactly_lib.section_document.element_parsers.token_stream_parser import TokenParser
+from exactly_lib.section_document.element_parsers.token_stream_parser import TokenParser, ParserFromTokens
 from exactly_lib.tcfs.path_relativity import RelOptionType
 from exactly_lib.type_val_deps.types.list_ import list_sdvs
 from exactly_lib.type_val_deps.types.path import rel_opts_configuration
@@ -26,13 +28,28 @@ REL_OPT_ARG_CONF = RelOptionArgumentConfiguration(REL_OPTIONS_CONF,
 
 
 def parser() -> Parser[ArgumentsSdv]:
-    return _PARSE
+    return _PARSER
 
 
 class _Parser(ParserFromTokenParserBase[ArgumentsSdv]):
     def __init__(self):
         super().__init__(consume_last_line_if_is_at_eol_after_parse=False,
                          consume_last_line_if_is_at_eof_after_parse=False)
+        self._elements_parser = generic_parser.ElementsUntilEndOfLineParser(
+            _ElementParser()
+        )
+
+    def parse_from_token_parser(self, token_parser: TokenParser) -> ArgumentsSdv:
+        elements = self._elements_parser.parse(token_parser)
+        return functools.reduce(_accumulate, elements, ArgumentsSdv.empty())
+
+
+def _accumulate(x: ArgumentsSdv, y: ArgumentsSdv) -> ArgumentsSdv:
+    return x.new_accumulated(y)
+
+
+class _ElementParser(ParserFromTokens[ArgumentsSdv]):
+    def __init__(self):
         self._element_choices = [
             parsing.TokenSyntaxSetup(
                 token_matchers.is_unquoted_and_equals(syntax_elements.REMAINING_PART_OF_CURRENT_LINE_AS_LITERAL_MARKER),
@@ -52,21 +69,7 @@ class _Parser(ParserFromTokenParserBase[ArgumentsSdv]):
             ),
         ]
 
-    def parse_from_token_parser(self, token_parser: TokenParser) -> ArgumentsSdv:
-        arguments = ArgumentsSdv.empty()
-
-        while not token_parser.is_at_eol:
-            following_arguments = self._parse_element(token_parser)
-            arguments = arguments.new_accumulated(following_arguments)
-
-        if self._consume_last_line_if_is_at_eol_after_parse:
-            token_parser.consume_current_line_as_string_of_remaining_part_of_current_line()
-        else:
-            token_parser.consume_remaining_part_of_current_line_as_string()
-
-        return arguments
-
-    def _parse_element(self, token_parser: TokenParser) -> ArgumentsSdv:
+    def parse(self, token_parser: TokenParser) -> ArgumentsSdv:
         return parsing.parse_mandatory_choice_with_default(token_parser,
                                                            syntax_elements.ARGUMENT_SYNTAX_ELEMENT_NAME.name,
                                                            self._element_choices,
@@ -99,10 +102,9 @@ def _parse_existing_path(token_parser: TokenParser) -> ArgumentsSdv:
     return arguments_sdvs.ref_to_path_that_must_exist(path)
 
 
-def _parse_plain_list_element(parser: TokenParser) -> ArgumentsSdv:
-    token = parser.consume_mandatory_token('Invalid list element')
-    element = parse_list.element_of(token)
+def _parse_plain_list_element(token_parser: TokenParser) -> ArgumentsSdv:
+    element = parse_list.ELEMENT_PARSER.parse(token_parser)
     return ArgumentsSdv.new_without_validation(list_sdvs.from_elements([element]))
 
 
-_PARSE = _Parser()
+_PARSER = _Parser()