+++ src/exactly_lib/section_document/defs.py
@@ -6,6 +6,7 @@ DESCRIPTION_DELIMITER_CHAR_NAME = name.NameWithGenderWithFormatting(
     name.a_name(name.name_with_plural_s('backtick'))
 )
 END_OF_FILE = 'END-OF-FILE'
+END_OF_LINE = 'END-OF-LINE'
 
 INSTRUCTION = name.NameWithGenderWithFormatting(
     name.an_name(name.name_with_plural_s('instruction'))
@@ -14,3 +15,7 @@ INSTRUCTION = name.NameWithGenderWithFormatting(
 INSTRUCTION_DESCRIPTION = name.NameWithGenderWithFormatting(
     name.an_name(name.name_with_plural_s('instruction description'))
 )
+
+
+def format_constant(constant: str) -> str:
+    return '"' + constant + '"'