+++ test/exactly_lib_test/impls/types/parse/z_package_suite.py
@@ -1,13 +1,12 @@
 import unittest
 
-from exactly_lib_test.impls.types.parse import symbol_syntax, misc_utils, parse_list
+from exactly_lib_test.impls.types.parse import symbol_syntax, misc_utils
 
 
 def suite() -> unittest.TestSuite:
     ret_val = unittest.TestSuite()
     ret_val.addTest(misc_utils.suite())
     ret_val.addTest(symbol_syntax.suite())
-    ret_val.addTest(parse_list.suite())
     return ret_val