+++ test/exactly_lib_test/impls/types/z_package_suite.py
@@ -9,6 +9,7 @@ from exactly_lib_test.impls.types.integer import z_package_suite as integer
 from exactly_lib_test.impls.types.integer_matcher import z_package_suite as integer_matcher
 from exactly_lib_test.impls.types.interval import z_package_suite as interval
 from exactly_lib_test.impls.types.line_matcher import z_package_suite as line_matcher
+from exactly_lib_test.impls.types.list_ import z_package_suite as list_
 from exactly_lib_test.impls.types.logic import z_package_suite as logic
 from exactly_lib_test.impls.types.matcher import z_package_suite as matcher
 from exactly_lib_test.impls.types.parse import z_package_suite as parse
@@ -27,9 +28,10 @@ def suite() -> unittest.TestSuite:
     ret_val.addTest(regex.suite())
     ret_val.addTest(expression.suite())
     ret_val.addTest(interval.suite())
+    ret_val.addTest(parse.suite())
     ret_val.addTest(string.suite())
     ret_val.addTest(path.suite())
-    ret_val.addTest(parse.suite())
+    ret_val.addTest(list_.suite())
     ret_val.addTest(logic.suite())
     ret_val.addTest(files_condition.suite())
     ret_val.addTest(matcher.suite())