+++ test/exactly_lib_test/test_resources/source/abstract_syntax_impls.py
@@ -102,7 +102,7 @@ class WithinParensAbsStx(AbstractSyntax):
             TokenSequence.sequence([layout.NEW_LINE, reserved_words.PAREN_END])
             if self._end_paren_on_separate_lines
             else
-            TokenSequence.singleton(reserved_words.PAREN_END)
+            TokenSequence.sequence([layout.OPTIONAL_NEW_LINE, reserved_words.PAREN_END])
         )