+++ test/exactly_lib_test/type_val_deps/types/program/test_resources/argument_abs_stxs.py
@@ -1,6 +1,7 @@
 import enum
 from typing import Optional, Sequence
 
+from exactly_lib.impls.types.list_ import parse_list
 from exactly_lib.impls.types.program import syntax_elements
 from exactly_lib.util.parse.token import QuoteType
 from exactly_lib_test.symbol.test_resources import token_sequences as symbol_tok_seq
@@ -39,6 +40,20 @@ class ArgumentOfStringAbsStx(ArgumentAbsStx):
 
     def tokenization(self) -> TokenSequence:
         return self.string.tokenization()
+
+
+class ContinuationTokenFollowedByArgumentAbsStx(ArgumentAbsStx):
+    CONTINUATION_TOKEN = parse_list.CONTINUATION_TOKEN
+
+    def __init__(self, argument_on_next_line: ArgumentAbsStx):
+        self._argument_on_next_line = argument_on_next_line
+
+    def tokenization(self) -> TokenSequence:
+        return TokenSequence.concat([
+            TokenSequence.singleton(parse_list.CONTINUATION_TOKEN),
+            TokenSequence.new_line(),
+            self._argument_on_next_line.tokenization(),
+        ])
 
 
 class ArgumentOfTextUntilEndOfLineAbsStx(ArgumentAbsStx):